p2p/discover: schedule revalidation also when all nodes are excluded (#30239)
If `nextTime` has passed, but all nodes are excluded, `get` would return
`nil` and `run` would therefore not invoke `schedule`. Then, we schedule
a timer for the past, as neither `nextTime` value has been updated. This
creates a busy loop, as the timer immediately returns.

With this PR, revalidation will be also rescheduled when all nodes are
excluded.

---------

Co-Authored-By: lightclient <lightclient@protonmail.com>