core/state: correct account selfdestruct EIP-6780 (part of 7702) (#53)
* pick up ethereum/go-ethereum#29520

* core/state: update test cases

* core/state: fix SetStorage override behavior (#30185)

This pull request fixes the broken feature where the entire storage set is overridden.

Originally, the storage set override was achieved by marking the associated account
as deleted, preventing access to the storage slot on disk. However, since #29520, this
flag is also checked when accessing the account, rendering the account unreachable.

A fix has been applied in this pull request, which re-creates a new state object with all
account metadata inherited.

* fixup! pick up ethereum/go-ethereum#29520

* fixup! core/state: fix SetStorage override behavior (#30185)

---------

Co-authored-by: rjl493456442 <garyrong0905@gmail.com>